Implemented: logic to store the transform mapping and then use those mappings(#2ftb86d)
